rename _pycmds to _main (#293)
* rename _pycmds to _main

* review
